new_project-sem5and6/requirements.txt
argon2-cffi>=23.1.0
cryptography>=42.0.0


